Split PatientFieldData into own file and fix RingdownForm bug
